# openapi.yaml
openapi: 3.0.3
info:
  title: SGS - Système de Gestion Scolaire API
  description: |
    API RESTful pour le Système de Gestion Scolaire (SGS).
    Cette spécification définit tous les endpoints nécessaires pour le backend.
    
    ## Authentification
    L'API utilise JWT Bearer tokens pour l'authentification.
    Tous les endpoints (sauf /auth/login) requièrent un header Authorization valide.
    
    ## Pagination
    Les endpoints de liste supportent la pagination avec les paramètres:
    - `page`: numéro de page (défaut: 1)
    - `per_page`: éléments par page (défaut: 20, max: 100)
    - `sort_by`: champ de tri
    - `sort_order`: asc ou desc
    - `search`: recherche textuelle
    
    ## Codes d'erreur
    - 400: Requête invalide
    - 401: Non authentifié
    - 403: Accès refusé
    - 404: Ressource non trouvée
    - 422: Erreur de validation
    - 500: Erreur serveur
  version: 1.0.0
  contact:
    name: Équipe SGS
    email: dev@sgs.edu
servers:
  - url: http://localhost:3001/api
    description: Serveur de développement
  - url: https://api.sgs.edu
    description: Serveur de production
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Connexion utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: admin@sgs.edu
                password:
                  type: string
                  format: password
                  example: Password123
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Auth]
      summary: Récupérer le profil utilisateur courant
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profil utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /comptabilite/depenses:
    get:
      tags: [Comptabilité]
      summary: Liste des dépenses
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/per_page'
      responses:
        '200':
          description: Liste paginée des dépenses
    post:
      tags: [Comptabilité]
      summary: Enregistrer une dépense
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepenseCreate'
      responses:
        '201':
          description: Dépense enregistrée
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    per_page:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
...
  responses:
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
            refresh_token:
              type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        nom:
          type: string
        prenom:
          type: string
        role:
          type: string
          enum: [super_admin, admin, comptable, rh, enseignant, parent, eleve]
        permissions:
          type: array
          items:
            type: object
            properties:
              module:
                type: string
              actions:
                type: array
                items:
                  type: string
        is_active:
          type: boolean
...
    DepenseCreate:
      type: object
      required: [libelle, categorie, montant, date_depense, mode_paiement]
      properties:
        libelle:
          type: string
        categorie:
          type: string
          enum: [fournitures, equipement, maintenance, salaires, services, autre]
        montant:
          type: number
        date_depense:
          type: string
          format: date
        beneficiaire:
          type: string
        mode_paiement:
          type: string
          enum: [especes, cheque, virement, mobile_money, carte]
        reference:
          type: string
...